Fetch restaurant reviews and save to IndexedDB
- In `openIDB` (in `dbhelper.js`), add a case 2 that opens a `reviews`
store in IDB
- In `dbhelper.js`, create a new `fetchReviewsByRestaurantId` that
fetches the reviews for a particular restaurant from the reviews
endpoint and saves them to the `reviews` store in IDB. The function
takes a restaurant ID and a callback function to execute.
- In `js/restaurant_info.js`, refactor the `fillRestaurantHTML` function
to run `DBHelper.fetchReviewsByRestaurantId`, which in turn calls
`fillReviewsHTML`, instead of running `fillReviewsHTML` directly.
- In `js/restaurant_info.js`, refactor `fillReviewsHTML` to accept
`error` and `reviews` arguments (one of which will be null) when called
from `DBHelper.fetchReviewsByRestaurantId`.
- TODO: Figure out why reviews stored in IDB are being successfully
returned while offline