Save new review to remote database while online
- Refactor url string in `DATABASE_REVIEWS_URL` to be more versatile,
and refactor `fetchReviewsByRestaurantId` which depends on it
- Create a new `DBHelper.saveReview` function that makes a POST request
to the reviews endpoint, passing in the review data received from
`validateFormData`, which now includes the restaurant ID.
- If POST is successful, pass review to new `displayNewReview` function
to be printed to page without a new fetch.
- TODO: This currently only works when online. Make it cache online
responses to IDB for later use and make it load reviews from IDB if it
can't reach the live server.
- TDOD: Currently refreshing the restaurant page doesn't make recently
added reviews show up. Figure out why and fix it.