Load reviews from IDB while offline
- Fix `sw.js` to prevent service worker from intercepting calls to fetch
reviews
- In `fetchReviewsByRestaurantId`, refactor to first look for reviews on
live server (and save to IndexedDB if found), then look in IndexedDB if
they can't be found online. If none are found there either, an error
message is passed to the callback function.
- In the callback function, `fillReviewsHTML`, make a distinction
between the case in which no reviews are found on the server and one in
which none are found in IndexedDB, and adjust messages displayed
accordingly.
- Refactor `restaurant.html` so the form will display above a list of
reviews but below an error message when there are no reviews.
- TODO: Save new reviews to IndexedDB so that they can display on page
refresh while offline. (Currently they are programmatically printed to
the page but not stored in IDB.)