Save new reviews to sync queue in IDB if offline
- In `dbhelper.js`, refactor `saveReview` function to first save the new
entry to the  `reviews` store in IDB by calling the new
`saveReviewInIDB` function,  regardless of network status, then attempt
to save to the network, then, if that fails, save to a new object in
store called `sync_queue` via the new `queueReviewInIDB` function. In
any case, the review is passed back to the callback function with a new
boolean `offlineStatus` argument representing whether the review is
queued offline.
- In `restaurant.js`, update `displayNewReview` and `createReviewHTML`
functions to take the new `offlineStatus` argument. In `createReviewHTML`
this is used to toggle on or off a class of `offline` applied to the `li`
for the review in question. The associated syle in `styles.css` indicates
offline status with a red border.
- TODO: Currently this status displays correctly if you never refresh
the page. However, if you refresh, the page will load from IDB or the
live server and all records will show as online.
- TODO: Sync items in the `sync_queue` to the server when connection is
re-established, then remove them from the queue
- TODO: Consider only saving to `reviews` object store the ones that
have been saved online, and pulling from both there and `sync_queue` when
loading records while offline